requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
dj-database-url==0.5.0
Django==3.1.3
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
psycopg2==2.8.6
pycodestyle==2.6.0
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
